requirements.txt
Flask==2.3.2
google-cloud-dialogflow-cx==1.25.0
gunicorn==20.1.0
google-auth==2.22.0
google-cloud-storage==2.11.0
google-api-python-client==2.118.0
google-cloud-datastore==2.19.0
python-dateutil==2.8.2
google-cloud-discoveryengine==0.11.7
google-api-core==2.17.1
oauth2client==4.1.3
